feat: Refactor routing configuration using createRoutesFromElements
The code changes refactor the routing configuration in the App.js file to use the createRoutesFromElements function from the react-router-dom library. This improves the readability and maintainability of the code by providing a more declarative way to define routes. The changes also update the router variable to use the new routesDefinition. This commit enhances the routing functionality of the application.